pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speedrun-wiki-sync"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pywikibot",
  "requests",
  "PyYAML",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
